particleCOntact & particleContactResolver avancés, particleContact attend la méthode de gestion de repos d'une particle. Les scripts des particleRod, wallContact ect sont à faire

+++ moteur_physique/moteur_physique/Sources/Forces/ParticleDrag.cpp
@@ -32,5 +32,5 @@ void ParticleDrag::updateForce(Particle* particle, float duration)
 	dragCoeff = Vector3D::get_normalization(velocity) * sum;
 
 	//On applique la force de trainée pendant la durée de la frame à notre vitesse. 
-	particle->setVelocity((velocity - dragCoeff) * duration);
+	particle->setVelocity((velocity - dragCoeff) * (double)duration);
 }
